==> Code/06-formatar-data.ipynb <==
# Resumo: Formatação de Datas em Repositórios Categorizados

Este script processa um arquivo CSV contendo repositórios categorizados, formata as colunas de datas (`created_at` e `updated_at`) no padrão `YYYY-MM-DD` e salva o resultado em um novo arquivo.

## Importação de Bibliotecas

- **pandas (`pd`)**: Para manipulação de dados em formato tabular.

## Estrutura do Código

1. **Importação de Bibliotecas**
   - Utiliza `pandas` para manipulação de dados.

2. **Carregamento de Dados**
   - O arquivo `repositorios_categorizados.csv` é carregado como um DataFrame.

3. **Formatação de Datas**
   - As colunas `created_at` e `updated_at` são convertidas para o formato de data `YYYY-MM-DD`.

4. **Exibição dos Dados**
   - O DataFrame atualizado é exibido no console.

5. **Exportação dos Resultados**
   - Os dados processados são salvos em um novo arquivo CSV (`repositories.csv`).


import pandas as pd

# Carregue a tabela
df = pd.read_csv('../Database/Resultados-Busca/repositorios_categorizados.csv')

# Formate as colunas "created_at" e "updated_at" como datas
df['created_at'] = pd.to_datetime(df['created_at'])
df['updated_at'] = pd.to_datetime(df['updated_at'])

# Formate as datas como "YYYY-MM-DD"
df['created_at'] = df['created_at'].dt.strftime('%Y-%m-%d')
df['updated_at'] = df['updated_at'].dt.strftime('%Y-%m-%d')

# Imprima a tabela formatada
print(df)

df.to_csv('../Database/Resultados-Busca/repositories.csv', index=False, sep=',', decimal=',')
