==> Code/08-busca-pelo-arquivos-config.ipynb <==
# Busca pelos arquivos de configurações dos repositórios

Na tentativa de responder a primeira questão de pesquisa, primeiro é feito a busca do arquivos de configurações do dataset gerado, como por exemplo o arquivo package.json e etc

## Importação de Bibliotecas

- **pandas (`pd`)**: Para manipulação de dados em formato tabular.
- **matplotlib.pyplot (`plt`)**: Para visualização de gráficos.
- **os**: Para interagir com o sistema operacional, como criação de diretórios.

## Funcionalidades

1. **Autenticação no GitHub**: O script começa com a autenticação na API do GitHub usando um token de acesso pessoal.

2. **Carregamento de Dados**: Carrega um arquivo CSV que contém uma lista de repositórios do GitHub.

3. **Leitura de Arquivos**: Define uma função `ler_arquivo` que tenta ler o conteúdo de arquivos específicos dentro de cada repositório. Se a leitura falhar, retorna `None`.

4. **Iteração sobre Repositórios**: Para cada repositório listado no CSV, o script:
   - Tenta acessar o repositório no GitHub.
   - Lê o conteúdo de uma lista predefinida de arquivos (ex: `package.json`, `vite.config.js`, etc.).
   - Armazena os resultados em uma lista.

5. **Salvamento dos Resultados**: Cria um DataFrame a partir dos resultados coletados e salva em um novo arquivo CSV, garantindo que o diretório de saída exista.


## Observações**
Esse script tem um tempo médio de 2 a 3 horas de execução.

import pandas as pd
from github import Github
import os

# Substitua 'seu_token_de_acesso' pelo seu token de acesso pessoal
token = "SEU-TOKEN"
g = Github(token)

# Carrega o arquivo CSV
# Substitua pelo caminho do seu arquivo CSV
csv_file_path = "../Database/Resultados-Busca/repositories.csv"
data = pd.read_csv(csv_file_path)

# Função para ler o conteúdo de um arquivo específico


def ler_arquivo(repo, caminho_arquivo):
    try:
        arquivo = repo.get_contents(caminho_arquivo)
        return arquivo.decoded_content.decode('utf-8')
    except Exception as e:
        # Retorna None se ocorrer um erro
        return None


# Lista para armazenar os dados
resultados = []

# Itera sobre cada repositório no DataFrame
for index, row in data.iterrows():
    repo_name = row['name']  # Nome do repositório
    print(f"Lendo repositório: {repo_name}")

    try:
        repo = g.get_repo(repo_name)
        arquivos_para_ler = [
            'package.json',
            'package-lock.json',
            'composer.json',
            'composer.lock',
            'vite.config.js',
            'angular.json',
            'tsconfig.json',
            'babel.config.js',
            'gatsby-config.js',
            'postcss.config.js',
            'vue.config.js'
        ]

        # Lê e armazena o conteúdo de cada arquivo
        for arquivo in arquivos_para_ler:
            conteudo = ler_arquivo(repo, arquivo)
            if conteudo is not None:  # Apenas adiciona se não houver erro
                resultados.append({
                    'repository': repo_name,
                    'file': arquivo,
                    'content': conteudo
                })

    except Exception as e:
        print(f"Erro ao acessar o repositório {repo_name}: {e}")

# Cria um DataFrame com os resultados
df_resultados = pd.DataFrame(resultados)

# Cria a pasta ../Database/pesquisa se não existir
output_dir = '../Database/Questao1'
os.makedirs(output_dir, exist_ok=True)

# Salva os resultados em um arquivo CSV
output_file_path = os.path.join(output_dir, 'resultados.csv')
df_resultados.to_csv(output_file_path, index=False)

print(f"Resultados salvos em: {output_file_path}")
